Move enterprise login form and tenant selection into a modal dialog
Extract login/password form and company selection into EnterpriseLoginModal,
keeping EnterpriseSettings as a simple connected/not-connected status view.
